Fixed #990: Permit passing additional parameters to SSH in "qvd-admin vm ssh".
This turned out to be trickier than I thought because GetOpt::Long removes "--" from @ARGV, so there's no way of knowing if the command was executed with parameters for SSH or with something completely different. As it stands, a command such as "qvd-admin vm ssh invalid" will not give an error; it will just add "invalid" to the SSH command line.

If you use the GetOpt::Long "pass_through" option "--" is kept in @ARGV, but, at least with the current version of GetOpt::Long, this means that options like -f and -q would have to come before the command, like in "qvd-admin -f user=foo vm list". Obviously this isn't what we want.



git-svn-id: https://svn.qindel.com/repos/main/projects/qindel/QVD/src/trunk@16337 4046b530-fbad-4e16-af16-ee11f0b55167